Start abstracting Trajectory
